remove unneeded JS-API dep move auth in the right place
