+++ lib/core/src/lib/auth/services/base-authentication.service.ts
@@ -15,14 +15,13 @@
  * limitations under the License.
  */
 
-import { PeopleApi, UserProfileApi, UserRepresentation } from '@alfresco/js-api';
 import { HttpHeaders } from '@angular/common/http';
-import { RedirectionModel } from '../auth/models/redirection.model';
-import { from, Observable, Observer, ReplaySubject, throwError } from 'rxjs';
-import { AppConfigService, AppConfigValues } from '../app-config/app-config.service';
-import { AlfrescoApiService } from './alfresco-api.service';
-import { CookieService } from '../common/services/cookie.service';
-import { LogService } from '../common/services/log.service';
+import { RedirectionModel } from '../models/redirection.model';
+import { Observable, Observer, ReplaySubject, throwError } from 'rxjs';
+import { AppConfigService, AppConfigValues } from '../../app-config/app-config.service';
+import { AlfrescoApiService } from '../../services/alfresco-api.service';
+import { CookieService } from '../../common/services/cookie.service';
+import { LogService } from '../../common/services/log.service';
 
 const REMEMBER_ME_COOKIE_KEY = 'ALFRESCO_REMEMBER_ME';
 const REMEMBER_ME_UNTIL = 1000 * 60 * 60 * 24 * 30;
@@ -33,18 +32,6 @@ export abstract class BaseAuthenticationService {
 
     onLogin = new ReplaySubject<any>(1);
     onLogout = new ReplaySubject<any>(1);
-
-    _peopleApi: PeopleApi;
-    get peopleApi(): PeopleApi {
-        this._peopleApi = this._peopleApi ?? new PeopleApi(this.alfrescoApi.getInstance());
-        return this._peopleApi;
-    }
-
-    _profileApi: UserProfileApi;
-    get profileApi(): UserProfileApi {
-        this._profileApi = this._profileApi ?? new UserProfileApi(this.alfrescoApi.getInstance());
-        return this._profileApi;
-    }
 
     constructor(
         protected alfrescoApi: AlfrescoApiService,
@@ -58,15 +45,11 @@ export abstract class BaseAuthenticationService {
     abstract isLoggedIn(): boolean;
     abstract isLoggedInWith(provider: string): boolean;
     abstract isOauth(): boolean;
-    abstract isImplicitFlow(): boolean;
-    abstract isAuthCodeFlow(): boolean;
     abstract login(username: string, password: string, rememberMe?: boolean): Observable<{ type: string; ticket: any }>;
     abstract ssoImplicitLogin(): void;
     abstract logout(): Observable<any>;
     abstract isEcmLoggedIn(): boolean;
     abstract isBpmLoggedIn(): boolean;
-    abstract getEcmUsername(): string;
-    abstract getBpmUsername(): string;
     abstract reset(): void;
     abstract once(event: string): Observable<any>;
 
@@ -190,15 +173,6 @@ export abstract class BaseAuthenticationService {
             return 'Basic ' + btoa(ticket);
         }
         return null;
-    }
-
-    /**
-     * Gets information about the user currently logged into APS.
-     *
-     * @returns User information
-     */
-    getBpmLoggedUser(): Observable<UserRepresentation> {
-        return from(this.profileApi.getProfile());
     }
 
     /**
